refactor: renamed Pagination component to PaginationButtons and moved the file
- Replaced instances of the Pagination component with the new PaginationButtons component across various content files to enhance navigation consistency.
- Updated imports and adjusted props to align with the new component structure.
- Removed the old Pagination component file to streamline the codebase.
