+++ islands/content/WalletDashboardContent.tsx
@@ -1,10 +1,10 @@
 /* ===== WALLET DASHBOARD CONTENT COMPONENT ===== */
 import type { DispenserRow as Dispenser, StampRow } from "$types/stamp.d.ts";
 
+import { PaginationButtons } from "$button";
 import { NOT_AVAILABLE_IMAGE } from "$constants";
 import { Icon, LoadingIcon } from "$icon";
 import { SortButton } from "$islands/button/SortButton.tsx";
-import { Pagination } from "$islands/datacontrol/Pagination.tsx";
 import { Setting } from "$islands/datacontrol/Setting.tsx";
 import FreshSRC20Gallery from "$islands/section/gallery/FreshSRC20Gallery.tsx";
 import { FreshStampGallery } from "$islands/section/gallery/FreshStampGallery.tsx";
@@ -188,7 +188,7 @@ function DispenserItem({
       {/* Pagination */}
       {pagination && pagination.totalPages > 1 && (
         <div class="mt-6">
-          <Pagination
+          <PaginationButtons
             page={pagination.page}
             totalPages={pagination.totalPages}
             prefix="dispensers"
