+++ routes/stamp/index.tsx
@@ -407,7 +407,7 @@ export function StampOverviewPage(props: StampPageProps) {
         pagination={{
           page,
           totalPages,
-          // Remove onPageChange to let Pagination component use its built-in Fresh navigation
+          // Remove onPageChange to let PaginationButtons component use its built-in Fresh navigation
         }}
       />
     </div>
